requirements.txt
joblib==1.5.2
pandas==2.3.3
streamlit==1.50.0
torch==2.9.0+cu126
pillow==12.0.0